Fixed typo's/Trigger conflict
Moved  triggerManager.removeTrigger("tempStoreOrgan"); to the beginning of organPlace(). Before it was canceling itself out before it could run the trigger. Now it runs the trigger, and on next check deletes itself and the new organtype can recreate it depending on the type of organ. Also there was a typo fixed where rusted blade was black trance.